Fix dependency mgmt in generated apps pom.xml
The previous commit for adding Tracing attempted to use dependency
management in the generated applications pom.xml but for some reason
the apps generator plugin always assumes the dep mgmt is for a POM
import and this fails at runtime.

This commit goes back to specifying the prometheus-rsocket-spring
with a version in the dependency.